services/etl/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4